Extend libArchetypeDefinition to include archetype_config
