--- processPolicySetDefinition_test.go
@@ -1,268 +1,0 @@
-package alzlib
-
-import (
-	"testing"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armpolicy"
-	"gotest.tools/v3/assert"
-)
-
-func Test_processSetPolicyDefinition_valid(t *testing.T) {
-	sampleData := getSamplePolicySetDefinition()
-	az := &AlzLib{
-		PolicySetDefinitions: make(map[string]*armpolicy.SetDefinition),
-	}
-
-	assert.NilError(t, processPolicySetDefinition(az, sampleData))
-	assert.Equal(t, len(az.PolicySetDefinitions), 1)
-	assert.Equal(t, *az.PolicySetDefinitions["Deploy-MDFC-Config"].Name, "Deploy-MDFC-Config")
-	assert.Equal(t, *az.PolicySetDefinitions["Deploy-MDFC-Config"].Properties.PolicyType, armpolicy.PolicyTypeCustom)
-}
-
-func getSamplePolicySetDefinition() []byte {
-	return []byte(`{
-		"name": "Deploy-MDFC-Config",
-		"type": "Microsoft.Authorization/policySetDefinitions",
-		"apiVersion": "2021-06-01",
-		"scope": null,
-		"properties": {
-			"policyType": "Custom",
-			"displayName": "Deploy Microsoft Defender for Cloud configuration",
-			"description": "Deploy Microsoft Defender for Cloud configuration",
-			"metadata": {
-				"version": "3.0.0",
-				"category": "Security Center"
-			},
-			"parameters": {
-				"emailSecurityContact": {
-					"type": "string",
-					"metadata": {
-						"displayName": "Security contacts email address",
-						"description": "Provide email address for Microsoft Defender for Cloud contact details"
-					}
-				},
-				"logAnalytics": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Primary Log Analytics workspace",
-						"description": "Select Log Analytics workspace from dropdown list. If this workspace is outside of the scope of the assignment you must manually grant 'Log Analytics Contributor' permissions (or similar) to the policy assignment's principal ID.",
-						"strongType": "omsWorkspace"
-					}
-				},
-				"ascExportResourceGroupName": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Resource Group name for the export to Log Analytics workspace configuration",
-						"description": "The resource group name where the export to Log Analytics workspace configuration is created. If you enter a name for a resource group that doesn't exist, it'll be created in the subscription. Note that each resource group can only have one export to Log Analytics workspace configured."
-					}
-				},
-				"ascExportResourceGroupLocation": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Resource Group location for the export to Log Analytics workspace configuration",
-						"description": "The location where the resource group and the export to Log Analytics workspace configuration are created."
-					}
-				},
-				"enableAscForSql": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForSqlOnVm": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForDns": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForArm": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForOssDb": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForAppServices": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForKeyVault": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForStorage": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForContainers": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				},
-				"enableAscForServers": {
-					"type": "String",
-					"metadata": {
-						"displayName": "Effect",
-						"description": "Enable or disable the execution of the policy"
-					}
-				}
-			},
-			"policyDefinitions": [
-				{
-					"policyDefinitionReferenceId": "defenderForOssDb",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/44433aa3-7ec2-4002-93ea-65c65ff0310a",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForOssDb')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForVM",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/8e86a5b6-b9bd-49d1-8e21-4bb8a0862222",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForServers')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForSqlServerVirtualMachines",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/50ea7265-7d8c-429e-9a7d-ca1f410191c3",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForSqlOnVm')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForAppServices",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/b40e7bcd-a1e5-47fe-b9cf-2f534d0bfb7d",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForAppServices')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForStorageAccounts",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/74c30959-af11-47b3-9ed2-a26e03f427a3",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForStorage')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderforContainers",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/c9ddb292-b203-4738-aead-18e2716e858f",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForContainers')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForKeyVaults",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/1f725891-01c0-420a-9059-4fa46cb770b7",
-					"parameters": {
-						"Effect": {
-							"value": "[parameters('enableAscForKeyVault')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForDns",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/2370a3c1-4a25-4283-a91a-c9c1a145fb2f",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForDns')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForArm",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/b7021b2b-08fd-4dc0-9de7-3c6ece09faf9",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForArm')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "defenderForSqlPaas",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/b99b73e7-074b-4089-9395-b7236f094491",
-					"parameters": {
-						"effect": {
-							"value": "[parameters('enableAscForSql')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "securityEmailContact",
-					"policyDefinitionId": "${root_scope_resource_id}/providers/Microsoft.Authorization/policyDefinitions/Deploy-ASC-SecurityContacts",
-					"parameters": {
-						"emailSecurityContact": {
-							"value": "[parameters('emailSecurityContact')]"
-						}
-					},
-					"groupNames": []
-				},
-				{
-					"policyDefinitionReferenceId": "ascExport",
-					"policyDefinitionId": "/providers/Microsoft.Authorization/policyDefinitions/ffb6f416-7bd2-4488-8828-56585fef2be9",
-					"parameters": {
-						"resourceGroupName": {
-							"value": "[parameters('ascExportResourceGroupName')]"
-						},
-						"resourceGroupLocation": {
-							"value": "[parameters('ascExportResourceGroupLocation')]"
-						},
-						"workspaceResourceId": {
-							"value": "[parameters('logAnalytics')]"
-						}
-					},
-					"groupNames": []
-				}
-			],
-			"policyDefinitionGroups": null
-		}
-	}`)
-}
